#68 Movie classes to internal package and rise the version
