+++ download-maven-plugin/src/main/java/com/googlecode/download/maven/plugin/internal/Artifact.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.googlecode;
+package com.googlecode.download.maven.plugin.internal;
 
 import org.apache.maven.artifact.factory.ArtifactFactory;
 import org.apache.maven.artifact.metadata.ArtifactMetadataSource;
